+++ CHANGES.rst
@@ -9,9 +9,16 @@ Changelog
 
 - Display column description if provided in schema `field.description`.
   [gbastien, bleybaert]
+
 - Specify in README.rst that versions >= 1.3 are for Plone5+ and
   versions < 1.3 are for Plone4.
   [gbastien]
+
+- Usability change: add an (hidden) label inside the add commands
+  [keul]
+
+- Compatibility with Plone 5 modals/overlay from mockup
+  [keul]
 
 1.3.0.post1 (2018-07-16)
 ------------------------